исправления замечаний
