+++ js/intro/intro-screen.js
@@ -1,6 +1,6 @@
 import IntroView from './intro-view';
 import {changeView} from '../utils';
-import Application from '../main';
+import Application from '../application';
 
 class IntroScreen {
   constructor() {
