+++ js/view.js
@@ -1,9 +1,15 @@
-// import { getElementFromTemplate } from './util';
-
 export default class AbstractView {
 
   get template() {
     throw new Error(`You have to define template for view`);
+  }
+
+  get element() {
+    if (!this._element) {
+      this._element = this.render();
+      this.bind();
+    }
+    return this._element;
   }
 
   render() {
@@ -14,14 +20,6 @@ export default class AbstractView {
 
   bind() {
 
-  }
-
-  get element() {
-    if (!this._element) {
-      this._element = this.render();
-      this.bind();
-    }
-    return this._element;
   }
 
 }
